# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/member/list": {
      "get": {
        "operationId": "MemberController_list",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "levelId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdFrom",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdTo",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activeFrom",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activeTo",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "excludePlaceholders",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "会员列表（分页/关键词）",
        "tags": [
          "member"
        ]
      }
    },
...
    "/member/{id}": {
      "get": {
        "operationId": "MemberController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "获取会员详情",
        "tags": [
          "member"
        ]
      },
      "put": {
        "operationId": "MemberController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "更新会员资料",
        "tags": [
          "member"
        ]
      },
      "delete": {
        "operationId": "MemberController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "删除会员",
        "tags": [
          "member"
        ]
      }
    },
...
    "/vehicle/list": {
      "get": {
        "operationId": "VehicleController_adminList",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码（从1开始）",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键词（车牌/品牌/会员）",
            "schema": {
              "example": "川A",
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "description": "范围：仅会员车辆/全部",
            "schema": {
              "default": "member",
              "type": "string",
              "enum": [
                "member",
                "all"
              ]
            }
          },
          {
            "name": "guest",
            "required": false,
            "in": "query",
            "description": "兼容旧参数：guest=1 时仅返回游客车辆",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "车辆列表（管理员，分页/关键词）",
        "tags": [
          "vehicle"
        ]
      }
    },
    "/vehicle/member/{memberId}": {
      "get": {
        "operationId": "VehicleController_listByMember",
        "parameters": [
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "按会员ID查询车辆",
        "tags": [
          "vehicle"
        ]
      },
      "post": {
        "operationId": "VehicleController_createForMember",
        "parameters": [
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleCreateForMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "为会员新增车辆（管理员）",
        "tags": [
          "vehicle"
        ]
      }
    },
    "/vehicle/member/by-phone": {
      "post": {
        "operationId": "VehicleController_createForMemberByPhone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleCreateForMemberByPhoneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "按手机号为会员新增车辆（管理员）",
        "tags": [
          "vehicle"
        ]
      }
    },
    "/vehicle/{id}": {
      "put": {
        "operationId": "VehicleController_updateVehicle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "修改车辆信息",
        "tags": [
          "vehicle"
        ]
      },
      "delete": {
        "operationId": "VehicleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "删除车辆",
        "tags": [
          "vehicle"
        ]
      }
    },
    "/vehicle/{id}/set-default": {
      "post": {
        "operationId": "VehicleController_setDefault",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "设置默认车辆",
        "tags": [
          "vehicle"
        ]
      }
    },
    "/vehicle/search": {
      "get": {
        "operationId": "VehicleController_search",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "模糊搜索车牌关键字",
            "schema": {
              "example": "川A",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "返回条数限制",
            "schema": {
              "default": 15,
              "example": 15,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "模糊搜索车牌（管理端/队列）",
        "tags": [
          "vehicle"
        ]
      }
    },
    "/vehicle/guest/create": {
      "post": {
        "operationId": "VehicleController_createGuest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleGuestCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "创建游客车辆",
        "tags": [
          "vehicle"
        ]
      }
    },
    "/vehicle/{id}/bind-member/{memberId}": {
      "post": {
        "operationId": "VehicleController_bindMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "将游客车辆绑定到会员",
        "tags": [
          "vehicle"
        ]
      }
    },
    "/vehicle/{id}/rebind": {
      "post": {
        "operationId": "VehicleController_adminRebind",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "一键换绑车辆（管理员）",
        "tags": [
          "vehicle"
        ]
      }
    },
    "/vehicle/{id}/rebind-logs": {
      "get": {
        "operationId": "VehicleController_adminRebindLogs",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码（从1开始）",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "车辆改绑记录（管理员，分页）",
        "tags": [
          "vehicle"
        ]
      }
    },
...
    "/vehicle/me/create": {
      "post": {
        "operationId": "VehicleController_myCreate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleMyCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "新增我的车辆（会员端）",
        "tags": [
          "vehicle"
        ]
      }
    },
...
    "/address/member/{memberId}": {
      "get": {
        "operationId": "AddressController_listByMember",
        "parameters": [
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "按会员ID查询收货地址（管理员）",
        "tags": [
          "address"
        ]
      }
    },
...
    "/group/{id}/members/{memberId}": {
      "delete": {
        "operationId": "GroupMemberController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupMember"
        ]
      }
    },
...
    "/group/{id}/vehicles/{vehicleId}": {
      "delete": {
        "operationId": "GroupVehicleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "vehicleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupVehicle"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wash Club API",
    "description": "API for 一体化洗车门店管理平台",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateMemberDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "昵称（≤10字符）"
          },
          "phone": {
            "type": "string",
            "description": "手机号",
            "example": "13800138000"
          },
          "password": {
            "type": "string",
            "description": "密码（可选）"
          },
          "points": {
            "type": "number",
            "description": "积分（可选）",
            "example": 0
          },
          "balance": {
            "type": "number",
            "description": "余额（可选）",
            "example": 0
          },
          "levelId": {
            "type": "object",
            "description": "会员等级ID（可选）",
            "nullable": true
          },
          "categoryId": {
            "type": "object",
            "description": "会员分类ID（可选）",
            "nullable": true
          },
          "tagIds": {
            "description": "标签ID列表（可选）",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "avatarUrl": {
            "type": "object",
            "description": "头像URL（可选）",
            "nullable": true
          }
        }
      },
...
      "VehicleCreateForMemberDto": {
        "type": "object",
        "properties": {
          "plateNumber": {
            "type": "string",
            "description": "车牌号",
            "example": "川A12345"
          },
          "vin": {
            "type": "object",
            "description": "VIN（可选）",
            "example": "LFV2A21K9E0000001",
            "nullable": true
          },
          "brand": {
            "type": "object",
            "description": "品牌（中文名，可选）",
            "example": "大众",
            "nullable": true
          },
          "series": {
            "type": "object",
            "description": "车系（中文名，可选）",
            "example": "高尔夫",
            "nullable": true
          },
          "brandId": {
            "type": "object",
            "description": "车型品牌ID（第三方库，可选）",
            "example": 123,
            "nullable": true
          },
          "seriesId": {
            "type": "object",
            "description": "车型车系ID（第三方库，可选）",
            "example": 456,
            "nullable": true
          },
          "typeMain": {
            "type": "string",
            "description": "车辆主类",
            "example": "SUV"
          },
          "typeSub": {
            "type": "object",
            "description": "车辆子类（可选）",
            "example": "紧凑型SUV",
            "nullable": true
          },
          "color": {
            "type": "object",
            "description": "车辆颜色（可选）",
            "example": "黑色",
            "nullable": true
          },
          "isDefault": {
            "type": "boolean",
            "description": "是否设为默认车辆",
            "example": false
          }
        },
        "required": [
          "plateNumber",
          "typeMain"
        ]
      },
      "VehicleCreateForMemberByPhoneDto": {
        "type": "object",
        "properties": {
          "plateNumber": {
            "type": "string",
            "description": "车牌号",
            "example": "川A12345"
          },
          "vin": {
            "type": "object",
            "description": "VIN（可选）",
            "example": "LFV2A21K9E0000001",
            "nullable": true
          },
          "brand": {
            "type": "object",
            "description": "品牌（中文名，可选）",
            "example": "大众",
            "nullable": true
          },
          "series": {
            "type": "object",
            "description": "车系（中文名，可选）",
            "example": "高尔夫",
            "nullable": true
          },
          "brandId": {
            "type": "object",
            "description": "车型品牌ID（第三方库，可选）",
            "example": 123,
            "nullable": true
          },
          "seriesId": {
            "type": "object",
            "description": "车型车系ID（第三方库，可选）",
            "example": 456,
            "nullable": true
          },
          "typeMain": {
            "type": "string",
            "description": "车辆主类",
            "example": "SUV"
          },
          "typeSub": {
            "type": "object",
            "description": "车辆子类（可选）",
            "example": "紧凑型SUV",
            "nullable": true
          },
          "color": {
            "type": "object",
            "description": "车辆颜色（可选）",
            "example": "黑色",
            "nullable": true
          },
          "isDefault": {
            "type": "boolean",
            "description": "是否设为默认车辆",
            "example": false
          },
          "phone": {
            "type": "string",
            "description": "会员手机号",
            "example": "13800138000"
          }
        },
        "required": [
          "plateNumber",
          "typeMain",
          "phone"
        ]
      },
      "VehicleUpdateDto": {
        "type": "object",
        "properties": {
          "plateNumber": {
            "type": "string",
            "description": "车牌号",
            "example": "川A12345"
          },
          "vin": {
            "type": "object",
            "description": "VIN（可选）",
            "example": "LFV2A21K9E0000001",
            "nullable": true
          },
          "brand": {
            "type": "object",
            "description": "品牌（中文名，可选）",
            "example": "大众",
            "nullable": true
          },
          "series": {
            "type": "object",
            "description": "车系（中文名，可选）",
            "example": "高尔夫",
            "nullable": true
          },
          "brandId": {
            "type": "object",
            "description": "车型品牌ID（第三方库，可选）",
            "example": 123,
            "nullable": true
          },
          "seriesId": {
            "type": "object",
            "description": "车型车系ID（第三方库，可选）",
            "example": 456,
            "nullable": true
          },
          "typeMain": {
            "type": "string",
            "description": "车辆主类",
            "example": "SUV"
          },
          "typeSub": {
            "type": "object",
            "description": "车辆子类（可选）",
            "example": "紧凑型SUV",
            "nullable": true
          },
          "color": {
            "type": "object",
            "description": "车辆颜色（可选）",
            "example": "黑色",
            "nullable": true
          },
          "isDefault": {
            "type": "boolean",
            "description": "是否设为默认车辆",
            "example": false
          }
        }
      },
      "VehicleGuestCreateDto": {
        "type": "object",
        "properties": {
          "plateNumber": {
            "type": "string",
            "description": "车牌号",
            "example": "川A12345"
          },
          "vin": {
            "type": "object",
            "description": "VIN（可选）",
            "example": "LFV2A21K9E0000001",
            "nullable": true
          },
          "brand": {
            "type": "object",
            "description": "品牌（中文名，可选）",
            "example": "大众",
            "nullable": true
          },
          "series": {
            "type": "object",
            "description": "车系（中文名，可选）",
            "example": "高尔夫",
            "nullable": true
          },
          "typeMain": {
            "type": "string",
            "description": "车辆主类（可选）",
            "example": "SUV"
          },
          "typeSub": {
            "type": "object",
            "description": "车辆子类（可选）",
            "example": "紧凑型SUV",
            "nullable": true
          },
          "color": {
            "type": "object",
            "description": "车辆颜色（可选）",
            "example": "黑色",
            "nullable": true
          }
        },
        "required": [
          "plateNumber"
        ]
      },
      "VehicleMyCreateDto": {
        "type": "object",
        "properties": {
          "plateNumber": {
            "type": "string",
            "description": "车牌号",
            "example": "川A12345"
          },
          "vin": {
            "type": "object",
            "description": "VIN（可选）",
            "example": "LFV2A21K9E0000001",
            "nullable": true
          },
          "brand": {
            "type": "object",
            "description": "品牌（中文名，可选）",
            "example": "大众",
            "nullable": true
          },
          "series": {
            "type": "object",
            "description": "车系（中文名，可选）",
            "example": "高尔夫",
            "nullable": true
          },
          "brandId": {
            "type": "object",
            "description": "车型品牌ID（第三方库，可选）",
            "example": 123,
            "nullable": true
          },
          "seriesId": {
            "type": "object",
            "description": "车型车系ID（第三方库，可选）",
            "example": 456,
            "nullable": true
          },
          "typeMain": {
            "type": "string",
            "description": "车辆主类",
            "example": "SUV"
          },
          "typeSub": {
            "type": "object",
            "description": "车辆子类（可选）",
            "example": "紧凑型SUV",
            "nullable": true
          },
          "color": {
            "type": "object",
            "description": "车辆颜色（可选）",
            "example": "黑色",
            "nullable": true
          },
          "isDefault": {
            "type": "boolean",
            "description": "是否设为默认车辆",
            "example": false
          }
        },
        "required": [
          "plateNumber",
          "typeMain"
        ]
      },
...
    }
  }
}